🔀 Redone blockState Map
Altered the blockState map to hold a list of emoji IDs instead of a HashMap with data IDs to associate with each emoji ID, essentially removed a layer from retrieving block state information. Additionally, new material types were added to the block constructor to allow for the changes to the blockState map.